Add files via upload
When we created this new repository, our hidden .gitignore file was not brought over. Adding it as it is needed for this project and there is a video step I need to show where the user checks to ensure this file exists and that our creds folder is not part of the repository.